First commit.
Created a sample HTML page with a form, and my semantic-form.css file, a
slightly modified version of what I have on SBD currently. The form is not
complete yet and I need to go through the CSS line-by-line and make sure
everything is as perfect as I can get it.